+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@dreamit/graphql-server",
-  "version": "4.4.0",
+  "version": "4.5.0",
   "description": "A GraphQL server written in NodeJS/Typescript.",
   "main": "build/src/index.js",
   "exports": {
